cleanup, including not crashing if we try to replace a decorated function
We ast.parse(func_source.strip()) to try to get the function's defaults.  We .strip()
so leading indentation isn't a problem.  However, if the function is decorated, we will
get a situation like "@decorator\n  def fn():..." and the parser will complain about
an invalid indent (following the decorator).  The fix for this for now is to put a
try...except around it, and just set defaults=[] if we can't parse the function.
This is bad because it means if we have a decorated function we will probably strip
the decorator *and* give the function the wrong types.  Decorators are essentially
just unsupported.