Make type disjointness less strict
Two classes A and B are considered disjoint unless A is in B's MRO or
vice versa. That means that we ignore some potential cases of
overlapping due to multiple inheritance.  The cases where this would
be useful are very rare.  Doing things the old way (i.e., strictly)
would complicate things for little gain.