Various updates to stubs related to co-/contravariance
We can't use covariant type variables in method argument types.
Removed such uses. Usually this is not too difficult.

Made Mapping invariant, because we can't currently give a
good signature for the get method if the value type is
covariant. The type context for the second argument to get
is wrong, resulting in type check failure when type checking
mypy. I'll add an issue for this.