Improve behavior of the type when merging branches.
For code of the form

  x = 'a' # type: Any
  if 1:
      x = 1
  else:
      x = 1

We now know that x has type int. If the last branch were x = 'a', then
x would have type Any instead.

This behavior is useful for weak mode, where everything has declared
type Any and we still want to do useful inference.  It may not be the
best behavior for non-weak mode, but I'm not sure.