Simplify many CallableType constructions that mostly copy
We often instantiate a new `CallableType` that's meant to
be just like an existing one with a few attributes changed.
This involves listing out all the attributes to copy over,
but that can obscure what the code is really trying to say,
and it's also error-prone, especially as new attributes
continue to be added over time.

Make a new method `copy_modified` to do this job in a more
explicit way, and switch a bunch of callsites to it.

As a side effect, fix a bunch of bugs (or at least latent
bugs) where we weren't copying one or more of the last few
attributes `bound_vars`, `line`, and especially
`is_ellipsis_args`.