Add new node for arguments.
Add a new node, consolidating the argument-type information from
FuncItem, which previously used parallel lists, into a list of
Argument.

Each argument contains:
  - The variable that it represents
  - The type annotation that was attached in the original AST. This
    is *not* the same as variable.type. (Hence why I named it something
    besides "type".
  - The initializer expression as a Node.
  - The initialization statement - this is the most similar to the old
    "init" list.

The initialization lvalue/statement are optional and are generated
by the initializer. (The only time they are passed in directly is by
the tree transformation phase, in order to preserve the lvalue's kind)