Add "-p" option to recursively typecheck package.
Adds a utility method in mypy.build that recurses a module. My
implementation actually feels a bit clunky so if anyone has better ideas
on how to do the traversal, that would be very welcome.

Uses os.cwdir() as the lib_path to find the modules themselves (because
we can't compute the lib_path that everything else uses yet,
unfortunately). I think this is somewhat reasonable.