Don't set breaking_out to False after visiting call expression.
I really don't know why I wrote that line in the first place, but it
seems completely wrong.  Removing it fixes bug #956.

This caused a test to fail.  The isinstance logic makes mypy skip
typechecking of dead code.  This change makes this more consistent,
causing the following to not be an error:

def foo() -> int:
    return foo()
    return 'a'

I changed the test to avoid this case.  We should probably decide what
we actually want to do about dead code and make appropriate tests.