Set breaking_out to False in try/except
We failed to set breaking_out to false in try/except, causing us to
not typecheck the rest of the function if the code raises an exception
or returns in the middle of one of the parts (#1040).

This commit also stops us from checking else_body twice.

Unlike in if/elif/else chains, we do not keep track of which parts of
the code path break out and which continue.  finally: blocks make this
tricky, and the main benefit was isinstance checks which don't happen
here.