Cache all searches for module through the path
This pushes the cache deeper so that it applies to every search of
this form.  In an empirical trial it doesn't seem to help much on its
own -- we do the actual search just about equally many times, and
spend equally much time doing so.

But the bulk of the work here, making an immutable lib_path which can
be used in a cache key, is shared with any other caching we might want
to do.  So this is useful preparation for the next commit which caches
some repeated work inside find_module and does cause a large speedup
empirically.