Cache repeated work inside find_module
Generally a program will import lots of modules that live
in the same packages as each other.  This means a lot of repeated
work looking for the directories where those packages would have
to live, and in a long lib_path most elements won't have most of
those directories.  Cache that work.