Optimize the string/path manipulation in find_module
This cuts out about 35% of the remaining time spent in find_module in
my testing, or about 0.7% of the total time spent by mypy.  That takes
find_module down to 1.4% of total time, from 10% before this and the
previous commit which cached subdirectory lookups.

Wouldn't be worth doing a lot of work to track this down and optimize,
but I've just studied this code anyway to do the caching, so it's
easy.  And I think the gain is worth the small complication of the
source code.