Add a line-count report
This report is aimed at measuring the total adoption of type
annotations in the codebase being type-checked.  The bottom-line (*)
output is the "number of lines of annotated code, as physical LOC."

(*) Metaphorically.  It's in the first line of the report file,
followed by the most significant individual files, for maximum
convenience in glancing at the file.

Because that isn't a definition that has a clear exact meaning in edge
cases (is an import statement annotated?), we make it precise in the
following way:
 * If all the functions in a file are annotated with signatures,
   then 100% of the file's physical lines count toward the total.
 * If none of the functions in a file are, then none of its lines do.
 * More generally, each function contributes to the total the value
   (# physical lines) * (# functions annotated) / (# functions total).

In principle a better formula might count up how big the annotated and
unannotated functions in a file are, to distinguish between a file
with some giant functions annotated and some trivial ones unannotated
and a file with vice versa.  In practice, at least at Dropbox so far
most files in a codebase that is beginning to adopt annotations are
close to either 0% or 100% annotated, so just about any way of
tweaking the edge cases would come to roughly the same result.