report: Include only the code specifically requested to typecheck
Before the advent of multiple sources in the CLI, there was no way for
us to distinguish the codebase the user was actually interested in
making sure is well-typed from the external dependencies that the user
cares about only to the degree necessary to type-check the codebase of
interest, because the only thing the user could do was to point to one
file and count on the relevant code, or most of it, coming in as
dependencies.

Now that we do accept multiple sources on the command line, the
natural way to type-check one's codebase is to explicitly direct the
type-checker at the whole thing, whether by directory tree, individual
filenames, module names, or packages.  So we can cut out noise from
external dependencies and keep the coverage reports focused on the
codebase of interest by limiting it to the given build sources.