report: Cut the use of "sources" entirely
Extracting whatever source happens to be first in the list is
kind of a silly operation in any case.  So just don't.
Originally the point of this was to identify the whole report
by the name of whatever single entry point we'd started from;
now that we no longer type-check from a single entry point,
just say directly that this file of the report is the index of
the whole thing.