runtests: Time how long we spend in different classes of tests
The measure here is "total task-elapsed time", which is kind of a
funny measure in the presence of concurrency and doesn't help with
identifying tasks that are extra costly because they're stragglers.
But to a first approximation the elapsed time of the whole test run
should be the total of this measure, divided by the number of tasks
we run concurrently, which is enough to make it useful for
identifying what's most profitable to work to cut.