runtests: typecheck self just once
A while ago we gained the ability to point mypy at a whole codebase
explicitly and typecheck all of it, rather than start from a single
file as an entry point and count on mypy going on to its dependencies
and other transitive dependencies.  That allows us to get the whole
thing in a single typecheck run, and know we've gotten it all, rather
than start from each file in the codebase in turn to make sure that we
don't miss one in case no one entry point has everything as transitive
dependencies.  Huzzah for something more like linear (*) time than
quadratic time.

(*) Empirically our type-checking runtime seems to be roughly linear
in the size of a codebase, and it'll always be important to maintain
that, though it might not be guaranteed in the worst case in principle.

On my desktop, this cuts about 4 minutes of total work from the test
run, spread across the various tasks running in parallel.