runtests: start eval-test first because it's the long pole
Running with 8-way concurrency on my desktop, and after the
optimizations made in the last few commits, the whole run finishes
when this one test does, long after all other tests have completed.
Clearly it ought to be broken up; first and easier, do the best we can
by putting it right at the head of the queue.