add Gentoo path to default_data_dir
Gentoo uses a wrapper around pythons to handle their dynamic default
python selection on a system with many versions of python installed
simultaneously.  Due to this mypy's script is a symlink to a file in
/usr/lib/python-exec/python3.5.  This means that the search path needs
to check if python-exec is being used and return the appropriate path
(usually '/usr/lib').  This patch works successfully on my Gentoo system.

Fixes #653.