Suppress line number for "ancestor" messages. Also drop a redundant `if`.
Suppressing the line number was David's idea. It's easy. But it has
one downside: in Emacs' grep/compile mode, you can't click on the line
to go to the file. Maybe that's okay, I have to think about it.

See the conversation in the PR for why my other idea (using the import
line as the context) won't work.