Make argument types in function type inference explicitly Optional
The old version happened to work because an argument would only be
omitted when the corresponding formal argument has a Callable type,
and in that case constraint generation does isinstance checks on the
actual argument type. But calling infer_constraints with actual=None
is not really good.

(I noticed this while experimenting with moving some Union argument
types to the second pass of function type inference.)