Invalidate incremental mode cache based on build flags
Let's have an example file:

% cat file.py
def f():
    pass
%

Before this patch the following would happen:

% mypy --incremental --disallow-untyped-defs file.py
file.py: note: In function "f":
file.py:1: error: Function is missing a type annotation
% mypy --incremental file.py
% mypy --incremental --disallow-untyped-defs file.py
%

With this patch applied mypy correctly notices when important flags change
between runs so it knows to ignore old cache.

The selection of flags is based on a quick look at what build flags are
available - it's likely I don't fully understand the consequences of
various flags and the selection will need to change.

I'm not entirely comfortable pushing a fix like this without an
automated test but I'm not sure how to approach testing this particular
behaviour as I'm not familiar with the codebase.