Make waiter cross-platform
Waiter depended on a couple of posix tricks. The first was a clever
use of os.waitpid(-1, 0) to wait for any spawned processes to return.
On Windows, this raises an exception - Windows has no concept of a
process group.

The second was the use of NamedTempFile(). On anything but Windows,
a file can be open for writing and for reading at the same time.
This trick actually isn't necessary the way NamedTempFile is used
here. It exposes a file-like object pointing to a file already opened
in binary mode. All we have to do is seek and read from it.