Fixing Windows tests
Most of these fixes revolve around the path separator and the way
Windows handles files and locking. There is one bug fix in here -
build.write_cache() was using os.rename to replace a file, which
fails on Windows. I was only able to fix that for Python 3.3 and up.