Cut the attribute ClassDef.base_types
This seems to be always equivalent to `bases` on the corresponding
TypeInfo, so keep the information in just one place.

This is part of clearing out the old branches I've had lying around
on my local clone of the repo -- back in December I was studying how
we handle the MRO, found this confusing, and had a note that this
attribute seemed to be redundant.  Further investigation bears that
out: in particular, the fact that the tests pass after this change
which cuts it out entirely.