Slightly clarify logic in analyzing base classes (#1696)
This is (a finished version of) another change/question I've had lying
around in my local clone for a few months.  It worried me that this
if/elif chain didn't end in an `else` -- if we fall off the end, surely
that's a bug either in the code we're typechecking or in mypy itself, no?
Or if somehow it's OK, then in any event that should be explicit here.
So put an `else` case on it, and while we're here combine that chain
and the TupleType conditional above it into one, reducing the number of
potential control flows for the reader to consider.

Part of understanding this code turned out to be realizing that
`base.fallback` is an `Instance` -- so make that explicit by putting
an explicit type annotation on the `base_types` local.  This seems to
be one of those occasions where the doctrine of "annotate function
signatures and wherever it's not obvious" counsels annotating a local
variable even though mypy is capable of inferring the type.

Two small semantic changes, both in error cases: when we can't make
sense of a base class, log an error but then treat it the same as we
would a base of Any and move on to the other base classes.  (In one
case we didn't report anything here, and in the other we panicked
harder.)  This regularizes our logic -- one fewer edge case where we
leave `info.mro` empty -- and should let us move on without reporting
extraneous errors down the line.  In the new `else` case, this changes
which error messages we emit for one test case, so update that test.

Also put each of the branches of our new unified if/elif/else chain in
the same form to make them a bit easier to follow and compare: first
report errors, if any, then record our conclusions for what it means
about the types.