Fix bug where
This commit fixes an additional bug with how mypy handles Type[T].
Consider the following code:

    from typing import Type, overload

    class Foo: pass

    @overload
    def f(a: Type[Foo]) -> str: None
    @overload
    def f(a: int) -> int: None

The expected behavior is for mypy to consider this code perfectly sound,
since Type[Foo] is not related to int. However, currently, mypy will
instead throw the following error:

    Overloaded function signatures 1 and 2 overlap with incompatible
    return types

This commit fixes this error.