Make TraverserVisitor not generic
None of the existing uses of TraverserVisitor use the ability to
return a value from the visit_* methods, and it's hard to imagine how
one could use that ability without overriding almost all of the
methods of TraverserVisitor. And the old code would be totally wrong
under strict optional checking.