Make Type[T] work with overloaded functions (#1770)
* Make Type[T] work with overloaded functions

Previously, Type[T] did not seem to respect overloaded functions. For
example, in the following snippit of code:

    from typing import Type, cast

    class Foo: pass

    print(str(cast(Type[Foo], Foo)))

...mypy would report the following error:

    error: No overload variant of "str" matches argument types [Type[test.Foo]]

...because str was defined to have the following type signatures:

    class str(Sequence[str]):
        @overload
        def __init__(self) -> None: ...
        @overload
        def __init__(self, o: object) -> None: ...
        @overload
        def __init__(self, o: bytes, encoding: str = None, errors: str = 'strict') -> None: ...

I previously thought the issue was that mypy did not recognize that
Type[T] was a subclass of object, but the problem was actually just that
the code to check the type signatures of overloaded functions was not
updated to work with Type[T].

This commit fixes an additional bug with how mypy handles Type[T].
Consider the following code:

    from typing import Type, overload

    class Foo: pass

    @overload
    def f(a: Type[Foo]) -> str: None
    @overload
    def f(a: int) -> int: None

The expected behavior is for mypy to consider this code perfectly sound,
since Type[Foo] is not related to int. However, currently, mypy will
instead throw the following error:

    Overloaded function signatures 1 and 2 overlap with incompatible
    return types

This commit fixes this error.