Add test for Options; fix bugs (#1789)
This PR adds a test that ensures the Options class and our argparse command line parser are in sync. In particular, it makes sure that after parsing the defaults are the same and that no new members are added to the Options object (which likely indicates a misspelling or forgetting to properly set dest).

This test catches a few missed dests (which I recall writing, but perhaps got lost in fixing merge conflicts), which are fixed here. It also fixes the lack of metavar on some special-options arguments, which caused weird looking help output.