Replace hack with a more principled check
After talking with Guido, I realized Python special-cases literally the
operators (==, +, etc), NOT the magic methods themselves. So, that means
we need to special-case "a == b", but not "a.__eq__(b)".

As a result, checking to see if the method name was a magic method or
not was the wrong thing to do -- if the name of the method is "__eq__",
there's no way to know if the code was originally "foo == bar" or
"foo.__eq__(bar)".

So, rather then trying to do something with the operator name, I can
instead check and see if the node is an OpExpr or ComparisonExpr node vs
a CallExpr node.