Fix import parent reference bug
This commit fixes the bug described in https://github.com/python/mypy/issues/1774.
The root issue ended up having to do with import statements of the
form "import a.b" and how dependencies were stored and retrieved in
incremental mode.

When the line "import a.b" is encountered, Python will actually, to some
extent, load module 'a' as a dependency so that dotted name lookups will
succeed. Mypy did not fully understand this, and considered only 'a.b`
as a dependency.

This commit will make mypy instead load both 'a' and 'a.b' to make the
behavior more consistent.

How mypy handles imports of the form "from a import b" is unchanged.