Add script to stress-test incremental mode
The `incremental_checker.py` script, when run within the `misc` folder,
will run incremental mode on a portion of mypy's commit history using
the version of mypy in the parent folder.

The checker script will check out an independent copy of the mypy repo
to test to avoid interfering with the current repo.