Make semantic analyzer mark ignored module imports
Previously, the semantic analyzer seemed to always mark import symbols as
being of kind MODULE_REF during the third pass and did not preserve
whether the module was previously marked with the `# type: ignore`
comment or not.

This caused a contradiction in some cases when mypy was run in incremental
mode: suppose file a.py imported file b.py, and marked that import with the
ignore annotation.

Since the import is being ignored, mypy would not register that import
as a dependency, but since the import (during the previous run) was
marked as a MODULE_REF, mypy would attempt to load that module anyways
in some cases.

This commit fixes https://github.com/python/mypy/issues/1904.