Include git hash in version.py during setup
This commit modifies setup.py to automatically edit the installed
mypy/version.py file during installation to include the git hash if
applicable.

If setup.py is not being run from a git repo, or if a git executable
can't be found in the system, mypy will default to using the original
version id.