Reset cache on mypy change (#1952)
* Make modify invalidate cache if version id changes

This commit modifies build.py to add an extra "version_id" attribute to
the cache metadata. If the version id specified in mypy/version.py
differs from the id stored in the cached data, mypy will automatically
invalidate that cached data.

* Include git hash in version.py during setup

This commit modifies setup.py to automatically edit the installed
mypy/version.py file during installation to include the git hash if
applicable.

If setup.py is not being run from a git repo, or if a git executable
can't be found in the system, mypy will default to using the original
version id.