Add a best-effort public interface extractor
"modinterface.py" is a new stage of the parsing process that ideally
happens relatively early, right after each file is parsed. It performs a
lightweight, best-effort-basis crawl of the parse tree to try and obtain
approximate type signatures of a given module to determine what its
"public interface" is.

If it cannot deduce the type signature of a given symbol, it resorts to
using a serialized form of the entire expression so that absolutely any
changes to how that symbol is derived is reflected.