Add "rechecked" keyword to incremental tests
This commit makes the incremental tests distinguish between "stale" and
"rechecked" files.

Stale files are ones where the "public interface" of that method is
changed; rechecked files are ones that must be re-analyzed (perhaps
because one of their dependencies changed?) but did not necessarily have
a changed public interface.

This functionality is currently unused, but paves a path for upcoming
incremental-mode changes.