Make incremental checks handle remote error
Previously, the incremental checks did not correctly handle the case where...

1. Module "A" imports and uses module "B"
2. Incremental mode is successfully run
3. Module "B" is changed such that module "B" typechecks, but module
   "A" now throws an error
4. Incremental mode is run again

What happens in this case is that the cache files for "A" and "B" are
produced in the first run, and only the cache files for "B" is updated
in the second run.

Previously, the checks made the assumption that if a file has an error,
it was due to a change within that file. This is not actually always the
case.