Manually set mtime instead of sleeping
Previously, the test suites for incremental added a delay of 0.1 seconds
between runs so that the mtime between runs would be different and a
file change would actually register with mypy.

Unfortunately, this doesn't necessarily work on some systems that have a
coarser mtime resolution.

To circumvent this, this commit gets rid of the sleep and just manually
adjusts the mtime upwards by a second so a file change is always
registered as a file change.