Fix for #2017
This commit fixes the issue in #2017. This bug manifests when...

1. You run mypy with silent_imports and incremental enabled against
   a subset of files
2. You re-run mypy with silent_imports and incremental enabled against
   a _slightly different_ subset of files (and end up needing to check
   a previously suppressed file in the process).

The issue appears to be that modules added to the "suppressed" field in
the cache metadata were assumed to also be suppressed in future runs of
mypy. Since the user can change what files they want to recheck, this
was an unsafe assumption.

This commit modifies the graph loading process so that any modules
previously marked as suppressed are manually rechecked again, just in
case.