Make incremental not propagate staleness if the public interface of a module is unchanged (v2) (#2014)
* Add tests for interface-based incremental checks

This commit modifies the incremental tests so that changes that don't
change the "public interface" of the file will not trigger a recheck in
any dependees.

* Add "rechecked" keyword to incremental tests

This commit makes the incremental tests distinguish between "stale" and
"rechecked" files.

Stale files are ones where the "public interface" of that method is
changed; rechecked files are ones that must be re-analyzed (perhaps
because one of their dependencies changed?) but did not necessarily have
a changed public interface.

This functionality is currently unused, but paves a path for upcoming
incremental-mode changes.

* Make incremental tests use 'rechecked' keyword

This commit updates the incremental tests to use the 'rechecked'
keywords and recalibrates a few tests to work against a more precise
interface change detector.

* Make incremental checks handle remote error

Previously, the incremental checks did not correctly handle the case where...

1. Module "A" imports and uses module "B"
2. Incremental mode is successfully run
3. Module "B" is changed such that module "B" typechecks, but module
   "A" now throws an error
4. Incremental mode is run again

What happens in this case is that the cache files for "A" and "B" are
produced in the first run, and only the cache files for "B" is updated
in the second run.

Previously, the checks made the assumption that if a file has an error,
it was due to a change within that file. This is not actually always the
case.