Make --py2 and --python-version mutually exclusive
This pull request will print a warning if the user attempts to pass in
both the --py2 and --python-version flags, since they are mutually
exclusive (argparse will pick whichever flag comes last).

I decided to make using both flags a warning instead of an error because
it may be potentially useful to override the Python version (for
example, when interacting with a script that wraps around mypy).