Record module name in TypeInfo
This commit modifies TypeInfo so it records the name of the module in
which the class (or NewType or namedtuple) was defined in. This does not
modify the behavior of mypy, but will make it easier to perform
introspection and analysis at later stages.

This information is also a part of the TypeInfo's fullname, but is
difficult to extract in the case of nested classes. For example, if we
have a fully qualified name of a.b.c.d, it's not always obvious whether
or not c is a module name or is a class name, and so forth.