Modify typechecker to record all module references
This commit modifies the typechecking stage to record all module
references that a particular module makes. This information will be
used in a future commit.

The motivation for this change is that it turns out that the set of
all module references is often larger then the set of explicitly
imported modules, and having this information would allow us to make
incremental mode more aggressive.