Remove the patch parent phase
This commit removes the "patch_parent" phase. The logic within those
phases are now run only when a file is actually importing something,
instead of being done automatically.

This fixes some irregularities with how imports are handled. In
particular, the "patch parent" stage was interacting poorly with
implicit module references since the automatic patch would add back in
an explicit module reference even when it was removed, which could
preserve an implicit reference that should have instead been broken.

That is, the presense of "patch parent" was sometimes concealing errors
that should have been caught.