Raise custom error when builtin symbol is missing
A common point of confusion people have when working on the mypy
codebase is forgetting to include the appropriate builtin fixture to
their unit tests. In particular, not including a fixture (or the right
fixture) can result in a cryptic exception when the code attempts to
look up the builtin symbol and fails.

This commit adds a check to the most common failure point and raises a
custom exception when it detects a builtin symbol is missing (which
should never happen when running mypy normally).