Keep track of if a Var is an import or not
This commit modifies Vars and adds an extra field to indicate if they
refer to imports or not. Normally, modules are not represented as Var
nodes, except in the case where that particular module was silenced or
suppressed in some manner.

This commit will make it easier to perform analysis on the AST during
the typechecking phase.