Add extra check to filter out non-modules
This commit adds an (ad-hoc) check to the indirect dependencies patcher
to filter out computed dependencies that are not, in fact, actually
modules. Ideally, this ad-hoc check shouldn't be necessary but there are
enough weird edge cases that it seems like a useful thing to have to
prevent unexpected behavior.