Allow errors on first run of incremental tests
This commit modifies the testing framework to allow us to specify the
expected output during both the first and second runs of incremental
mode, instead of mandating that the first run typechecks.

This is useful since it lets us test to make sure incremental recovers
correctly in the presence of errors.