Make cache files compressed by default
This commit modifies the cache writing process so that the JSON
generated for the cache files are not indented by default for
performance improvements.

This commit also adds a hidden `--debug-cache` flag. When the flag is
present, mypy will begin indenting the JSON cache data again to make it
easier for developers to debug.