Add `--disallow-subclassing-any` flag
This commit adds a new flag that disallows classes from having a base
class of type `Any`.

This tends to happen when using `--silent-imports` or when using a
class from a module or library that is not typechecked by mypy.
Normally, mypy will silently accept subclassing Any, which can sometimes
lead to discrepencies which further leads to subtle bugs.

Passing in this optional flag will raise an error instead.