Compress usually-false flags when serializing
This commit modifies several mypy nodes that tend to contain a large
number of usually-false boolean flags to store a compressed version of
those flags.

More precisely, rather then storing each flag and its boolean value as a
key-value pair in the output JSON, we now store a list of the flags that
are known to be true at time of serialization.

When deserializing, we iterate over the list of flags and set each
attribute to True, overriding the default value of False.

This does not result in any significant speedups, but does end up
reducing the size of the cache by a non-trivial factor.