Stop caching all arguments in serialized FuncDefs
This commit modifies the serialization logic and some of the surrounding
code for FuncDefs so we no longer serialize the full argument list --
instead, we just stick to serializing the argument names and the
argument kinds.

As with the other commits, this does not really improve speed, but does
end up shrinking the cache size by a fair amount.

It would have been nice to avoid storing the arg kinds and arg names
altogether since the information can usually be found within the
corresponding callable type, but unfortunately, a FuncDef is not
guaranteed to actually have a type in all instances -- sometimes, the
type is None. This is a problem since we need the argument names and
kinds to be able to infer a fallback type later on in mypy.