Allow identical classes to be repeated twice
This commit fixes https://github.com/python/mypy/issues/2135.

Previously, if you attempted importing two symbols of the same name
twice, mypy would check to see if that symbol was a variable definition
or a function and would allow the import if that symbol was previously
imported.

This commit extends this check slightly by also allowing identical
classes to be imported twice.