Fix overzealous new submodule check
Previously, when mypy detected a module acquired a new submodule after
running mypy once in incremental, mypy would attempt to parse the module
from scratch (if it hadn't already previously been parsed) which, as a
side-effect, would find all of the child modules and load them as
dependencies of the top-level module.

While this works in the usual case, this check made the flawed
assumption that we would want to actually record every single new
submodule that was created.

This assumption is not necessarily true when using --silent-imports --
it may be the case that we only want to record a _subset_ of the new
submodules that were created, depending on what exactly the command line
flags were.