Allowing importing a submodule ahead of its parent module.
Fixes #2133.

This is the third fix for this issue.  I think I've finally nailed it.
The solution is not to bail out early when the parent module is
unavailable -- if all the imports are submodules of a package that's
acceptable (the bug was that in incremental mode it's possible that
the parent is fresh and has not been entered into the modules dict
yet, but the bail-out caused the import to be marked as missing).