Add --warn-no-return, and major binder refactoring (#1748)
* Replace the binder's breaking_out by an 'unreachable' property of Frame

Also add some comments to the binder.

* Minor type cleanup in the binder and explanation of Key

This also fixes a bug where the index expression x['m'] was being treated
by the binder as equivalent to the member expression x.m.

* Add warn-no-return flag to give a note on missing return statements

* Learn that loop condition is False on exit from while loop

* Use bool() in tests when we want an unspecified bool in a conditional

* Handle 'while True', 'assert False'

* Reduce duplication with find_isinstance_check and binder.push

The resulting simplification of visit_if_expr revealed a bug in the
binder: if a frame was marked as unreachable, but then a new frame
context was begun immediately, the new frame was not marked as
unreachable, and a block visited inside the new frame would be type
checked. For now we check all frames for reachability; probably there
is a more efficient way to do this.

This commit also cleans up check_boolean_op slightly.

* Process try statements in top-to-bottom order

Fixes #1289 (again).

* Fix return type of binder.unreachable

* Skip docstring when determining whether function body is trivial

* Remove None return values from check_return_stmt

* Explain why two try frames are needed for a finally clause; add a test

* Decouple checker and binder by adding more options to frame_context

No more poking around by the checker in loop_frames and try_frames.