Don't learn anything from isinstance checks that always pass or fail
Previously in a case like

    class A: pass
    class B(A): pass
    x = B()
    if isinstance(x, A):
        ...

we would "learn" inside the `if` that x has type A (and forget that it has
type B), which is silly since B is a subtype of A anyway.

For symmetry we also learn nothing in the non-overlapping case, where an
isinstance check would always fail; this should make no difference since
the type we learned was the same as the current type.