Optionally override the origin context of an error (#2406)
`ErrorInfo` has a new `origin` attribute which is used as the origin
location of an error to determine whether it should be ignored.  The
regular context is used if `origin` is absent.  In any case, it is used
to display a location before a message, which may be different from the
origin.  Currently, the only such error is
`unexpected_keyword_argument`.

Fixes #2170