Add HTML report type with tooltips
This adds a HTML reporter that will produce mouse-over type tooltips in
the output.

The command line argument is --xslt-html-tooltips-report.

This also adds by-default-disabled functionality to
stats.StatisticsVisitor to build a detailed map of the form (Line,
Column) -> TypeString.

It also extends the report.MemoryXmlReporter to optionally add XML
elements from the extended StatisticsVisitor.