Change flag names for silent modes and add a new mode ("follow silently") (#2513)
There are now two flags:
```
  --ignore-missing-imports  silently ignore imports of missing modules
  --follow-imports {normal,silent,skip,error}
                            how to treat imports (default normal)
```
The first is pretty clear. The second works as follows (I still have to update the docs).

It only applies to .py modules (i.e. not to .pyi stubs) that are not specified on the command line but are needed to satisfy some import(s) and are found on the module search path (MYPYPATH plus the defaults derived from the command line).

The values mean the following:
- normal: process such modules normally (parse, analyze etc.)
- silent: process normally but silence (ignore) any non-blocker errors in those modules
- skip: silently don't process at all; the module is given type `Any`
- error: don't process at all and give a single error message for each such module

Note that when these flags are used as per-file flags, the pattern should match the imported module, not the importing module.