Fix performance degradation when you have more than 256 entries in mypy.ini (#2617)
Since we were calling fnmatch.fnmatch for every entry in mypy.ini for every file that gets checked performance would degrade sharply when the number of entries was more than 256.
That's because fnmatch.fnmatch lrucache 256 patterns internally. And because the access pattern is linear this would cause a perfect storm of (fnmatch.translate + re.compile) * entries * modules if you went over.

So to solve it we store the Pattern in the dict and use that to match.