Freshen type variables of generic methods (#2753)
This fixes conflicts where unrelated type variables were treated as
the same, because type variables defined in different scopes happened
to have the same ids. The fix is to substitute type variables with fresh
ones when looking up the types of methods. Fixes #2726.

Also fix method overlapping checks to unify type variables, as
otherwise the freshened generic methods may cause errors
elsewhere. Type variables don't need to be the same for signatures to
be compatible -- they just need to be equivalent.