Keep track of targets that generated an error
In fine-grained incremental mode keep reporting errors for targets
with errors even if the targets weren't explicitly triggered during
successive iterations. We currently repeatedly type check these
targets. Alternatively, we could perhaps just keep generating the
errors without actually re-type-checking the targets, but there's a
risk that it will be hard to always do this correctly, so we are
starting with this simple-minded approach which is good enough
for now.