fix stubgen crash on an extension module containing an alias of object
Closes https://github.com/python/typeshed/issues/766.

I verified that the new test case fails without the changes made in
stubgenc.py. Obviously a lot more tests for stubgenc could be added,
but that's for another day.

The original crash can be reproduced by creating file called alias.py
containing the line `alias = object`, compiling it to a .so with
Cython, and running stubgen on the .so. I got the following stack
trace:

$ python3 -m mypy.stubgen alias
Traceback (most recent call last):
  File ".../lib/python3.6/runpy.py", line 193, in _run_module_as_main
    "__main__", mod_spec)
  File ".../lib/python3.6/runpy.py", line 85, in _run_code
    exec(code, run_globals)
  File ".../lib/python3.6/site-packages/mypy/stubgen.py", line 734, in <module>
    main()
  File ".../lib/python3.6/site-packages/mypy/stubgen.py", line 632, in main
    raise e
  File ".../lib/python3.6/site-packages/mypy/stubgen.py", line 629, in main
    interpreter=options.interpreter)
  File ".../lib/python3.6/site-packages/mypy/stubgen.py", line 101, in generate_stub_for_module
    class_sigs=class_sigs)
  File ".../lib/python3.6/site-packages/mypy/stubgenc.py", line 41, in generate_stub_for_c_module
    generate_c_type_stub(module, name, obj, types, sigs=sigs, class_sigs=class_sigs)
  File ".../lib/python3.6/site-packages/mypy/stubgenc.py", line 169, in generate_c_type_stub
    if all_bases[-1] is object:
IndexError: list index out of range