"Quick and dirty" variant of incremental mode.
This uses the cache whenever the cache file exists and is newer than the source.

The difference with regular incremental mode is that regular
incremental mode forces a re-check whenever any dependency is stale,
and it forces re-checks for all modules in an SCC.  In quick-and-dirty
mode, by contrast, if some modules in an SCC have a new-enough cache
file, they are loaded from the cache even if some other module in the
same SCC or one of its dependencies was stale.

This is much quicker but obviously not always correct.  The hope is
that it speeds up checking single files when there are large SCCs (or
large dependency graphs) around.

A typical workflow would be:

1. mypy -i (slow, but warms the cache)
2. edit some file (or a few)
3. mypy --quick-and-dirty (re-checks only the file(s) you edited)
4. repeat steps 2-4 any number of times (maybe for different files)
5. mypy -i (re-check everything -- this could be part of CI)
6. if any errors left, go to step 2