Make Type[T] subtype of Callable[..., T] (#2894)
Fixes case like
```
from typing import Callable, Type
class A: pass
x: Callable[..., A]
y: Type[A]
x = y
```
Note that it shouldn't typecheck in the opposite direction. There is still problem with parameter types, as the comment says, but this isn't the problem I'm trying to fix.

This is needed for #2892, but isn't enough (we also need #2895).