Make it an error to use a class-attribute type var outside a type (#3105)
Previously, e9d28a01 fixed a crash when you tried to access a class-attribute
type variable. The test in that commit involved assigning a variable the value
of the typevar. It resulted in no crash, but rather treating the variable as
being an instance of the type the typevar bound to, later, which is incorrect.

Instead, this PR treats such an assignment as an error, and gives you the same
message as when you try to alias a typevar directly.

Also test a *correct* alias with the typevar access method in question -- it
works.