Use inheritance to allow most TypeVisitors not to deal with "synthetic" types (#3204)
* Use inheritance to allow most TypeVisitors not to deal with "synthetic" types

This is a nearly-complete fix to https://github.com/python/mypy/issues/730 --
the trick is that some types aren't real, they're synthetic AST constructs, and
so we shouldn't have to deal with those in every type visitor. Only the type
visitors that deal with types before the synthetic constructs are analyzed away.