Type check arguments for "super"
This is not quite as strict as we could be, but this is still better
than what we had before. A full implementation will be somewhat
complicated.

Remaining issues:

* Assume that the first argument refers to the enclosing class, even
  though that's not generally the case.

* Assume that a type object as the second argument is always fine.

* Don't validate some kinds of type object types as the first argument.

* Single-argument variant is not supported.

Fixes #526 (modulo the open issues).

Also, I no longer think that defining messages as constants in
`mypy.messages` is generally a good idea so I'm not following
that pattern here. I'll officially update the convention in a
separate PR.