Add test cases for AST serialization
The test cases use a serialized module in incremental mode
in various ways to make sure that all important information
in the AST nodes is preserved during serialization and
deserialization.

These are all very straightforward. The idea is to help catch
regressions early and to make it easier to refactor or optimize
serialization code. These found 2 issues -- separated the first
to PR #3275. The second issue (#3274) is still unfixed and there
is a skipped test case for it.

More complex cases such as import cycles will still go to
check-incremental.test.