myunit: Find temp directory relative to source tree, not cwd
We use a directory tmp-test-dirs/ under the top level of the mypy
source tree for our temporary directories when running tests.
We've been finding it relative to the current working directory, on
the assumption that that is the root of the source tree; which is a
fine constraint to require, but the way we fail if that constraint is
broken is an error in each test case's setup, which pytest presents as
one of those massive stack traces it specializes in.

Instead, find the root of the source tree directly, using `__file__`.

After this change, `pytest -n0` runs fine from `mypy/` or
`mypy/test/`, and plain `pytest` (with parallelization) runs fine from
`mypy/test/` and gives a different, now mercifully short, obscure
error from `mypy/`.  We'll leave that for another PR or another day.