Speed up tests by simplifying the stub for typing
Don't include async types in the default `typing` stub used in tests
since it's only needed for a small fraction of test cases.

Make it possible to write `[typing fixtures/typing-full.pyi]` in tests
to use a more complete complete stub that includes the async types.
In the future we can move additional rarely used features to this
stub.

Also replace the definition of `NewType` in the test fixtures with a
simpler one to speed things up slightly.

These changes improved the running time of `pytest mypy` by about 35%
for me (on macOS).