Fix spurious "cycle in function expansion" errors.
Most of these (at least) seem to be due to a function being analyzed
multiple times, and the 'expanded' list filling up with 2 or more
references to the same FuncDef.

If there are other reasons they will still trigger the same error, but
at least for mypy itself, the error no longer appears with this simple
fix.