Change reveal_type() representation of TypedDicts
Change the formatting to be closer to the TypedDict definition syntax
and preserve TypedDict name when constructing using TD(...) for better
reveal_type output.

Note that this uses a new representation `'key'?: type` for
non-required keys.

I'll update remaining error messages to use the new syntax in a
separate PR.

Fixes #3590.