Prohibit referring to class within its definition
This partially, but not completely, resolves
https://github.com/python/mypy/issues/3088

Previously, code of the following form was accepted by mypy:

    class A:
        def foo(self) -> A:
            ...

This results in an error at runtime because 'A' is not yet defined.
This pull request modifies the parsing process to report an error when
a type is being incorrectly used within its class definition.

However, this commit does *not* attempt to handle other cases where
the user is using a type that is undefined at runtime. For example, mypy
will continue to accept the following code without an error:

    def f() -> A:
        ...

    class A:
        ...

I decided it would probably be best to address this in a future pull
request.