Python 3.4 support (#3733)
The new typed_ast 1.1.0 includes a patch adding support for Python 3.4
and 3.3. The README is updated to discuss this. A comment on LXML on
Python 3.3/3.4 was added in case there was someone who really wanted to use reports. Also I removed the comment about the difficulty/easy tag on issues as that isn't used anymore.

3.3 ends up being supported incidentally as it uses the same compiler as 3.4 (thus typed_ast supports it, so we do).

lxml in test-requirements is not updated due to lxml not supplying wheels for Python 3.4 64 bit.