Implement compound operator assignments.
This required a surprising amount of refactor to make it possible
to sort of merge binary operations and assignment.

The fundamental problem is that `assign` expects the rvalue to
be an Expression and in the case of operator assignments we do
not have one available. We would actually rather assign directly
to the register using binary_op.

List index assignments should be fairly similar but it may be better
to implement after we have list of list support.

As a result of splitting up the "find assignment target" from "actually
do the assignment of an expression" we have a new abstraction which
I'm up for renaming... It should be the IR equivalent of an Lvalue
but I'm not sure what that would be.