Add support for tuples.
This had to reorganize a lot of the emitters code because the parts
outside of Emitter got too complicated and we needed to share code
between visit_box and the return result boxing and so forth.

NOTE: The tuples have basically nothing supported (indexing does
not work and no time to figure out why). However, the code generation
and unboxing/inner unboxing IS functional.