Erase typevars of generic types when used in Type[C]
Prior to this commit, when generic types are used as the C in a Type[C]
expression, they don't have their typevars substituted; this leads to
the internal type variables of their implementations leaking out in a
way that makes them difficult to use.

Instead, we now erase those typevars to Any.  This fixes #3824.