[Refactoring]Rename TypeOfAny.implicit to TypeOfAny.unannotated
TypeOfAny.unannotated designates Anys that come from omitted type annotations.

```
def f(x) -> None:
    pass
```
In the example above, type of `x` is `Any` and type of that `Any` is "implicit".

However, usage of the word "implicit" is ambiguous since you could consider
omitted generics, for instance, implicit as well but it is a different category
of Anys.