Infer better type for call to overload with Any actual arguments
If the overload return type is ambiguous and there is an `Any` actual
argument that seems to cause the ambiguity, infer `Any` as the type of
the call expression instead of some precise type. Previously we
sometimes inferred a precise type even though the type was ambiguous,
resulting in false positives.

This is not quite the most correct implementation since we sometimes
incorrectly assume ambiguity caused by `Any` even though none actually
exists. This is acceptable since this only affects calls with `Any`
arguments that are imprecise anyway. Also, the correct logic would be
quite complex, hard to test and probably not worth it, at least right
now.

Fixes #3662.